+++ api/countryData/restCountries.go
@@ -23,7 +23,7 @@ type MyError struct {
 }
 
 //Handler - Class function will be called and handle all requests and fetches
-func (data *Information) Handler(country string) (int, error) {
+func (data *Information) Handler(country string) (int, error, string) {
 
 	//Check if document exists, if it does not exist run this code:
 	if _, err := os.Stat("./data/countries.json"); os.IsNotExist(err) {
@@ -32,7 +32,7 @@ func (data *Information) Handler(country string) (int, error) {
 		var input Information
 		status, err := input.req("https://restcountries.eu/rest/v2/all")
 		if err != nil {
-			return status, err
+			return status, err, ""
 		}
 
 		//Store it in a file:
@@ -43,21 +43,21 @@ func (data *Information) Handler(country string) (int, error) {
 	//If function was called with a specific country
 	if country != "" {
 		var specificCountry Information
-		status, err := specificCountry.oneCountry(country)
+		status, err, countryCode := specificCountry.oneCountry(country)
 		if err != nil {
 
-			return status, err
+			return status, err, ""
 		}
-		return http.StatusOK, nil
+		return http.StatusOK, nil, countryCode
 
 	} else { //If function was NOT called with a specific country:
 		var AllData Information
 		status, err := AllData.allCountries()
 		if err != nil {
 
-			return status, err
+			return status, err, ""
 		}
-		return http.StatusOK, nil
+		return http.StatusOK, nil, ""
 	}
 }
 
@@ -77,29 +77,28 @@ func (data *Information) req(url string) (int, error) {
 }
 
 //oneCountry -Gets information about one specific country from local storage
-func (data *Information) oneCountry(countryName string) (int, error) {
+func (data *Information) oneCountry(countryName string) (int, error, string) {
 	countryName = strings.Title(strings.ToLower(countryName))
 	//Get data from document
 	file, err := ioutil.ReadFile("./data/countries.json")
 	if err != nil {
-		return http.StatusInternalServerError, err
+		return http.StatusInternalServerError, err, ""
 	}
 
 	err = json.Unmarshal([]byte(file), &data)
 	if err != nil {
-		return http.StatusInternalServerError, err
+		return http.StatusInternalServerError, err, ""
 	}
 
 	//Looping to look for the specific country
 	for _, v := range *data {
 		if v.Name == countryName {
-
 			//When found, break the loop and return that one instance
 			*data = Information{{v.Name, v.Alpha2Code, v.Capital}}
-			return http.StatusOK, nil
+			return http.StatusOK, nil, v.Alpha2Code
 		}
 	}
-	return http.StatusBadRequest, &MyError{"Country:" + countryName + " was not found in the database"}
+	return http.StatusBadRequest, &MyError{"Country:" + countryName + " was not found in the database"}, ""
 }
 
 //allCountries -Gets information about all countries from local storage