+++ api/weatherData/weatherData.go
@@ -1,17 +1,14 @@
 package weatherData
 
 import (
-	"errors"
 	"main/db"
 	"net/http"
+	"time"
 )
 
-// WeatherData structure stores current and predicted weather data for a day.
-//
-// Functionality: Handler, get
-type WeatherData struct {
-	Updated string `json:"updated"`
-	Now struct {
+// Timeseries stores current and predicted weather data for a day
+type Timeseries struct {
+	Instant struct {
 		AirTemperature      float64 `json:"air_temperature"`
 		CloudAreaFraction   float64 `json:"cloud_area_fraction"`
 		DewPointTemperature	float64 `json:"dew_point_temperature"`
@@ -20,8 +17,8 @@ type WeatherData struct {
 		WindSpeed           float64 `json:"wind_speed"`
 		WindSpeedOfGust     float64 `json:"wind_speed_of_gust"`
 		PrecipitationAmount float64 `json:"precipitation_amount"`
-	} `json:"now"`
-	Today struct {
+	} `json:"instant"`
+	Predicted struct {
 		Summary                     string  `json:"summary"`
 		Confidence                  string  `json:"confidence"`
 		AirTemperatureMax           float64 `json:"air_temperature_max"`
@@ -30,7 +27,15 @@ type WeatherData struct {
 		PrecipitationAmountMax      float64 `json:"precipitation_amount_max"`
 		PrecipitationAmountMin      float64 `json:"precipitation_amount_min"`
 		ProbabilityOfPrecipitation	float64 `json:"probability_of_precipitation"`
-	} `json:"today"`
+	} `json:"predicted"`
+}
+
+// WeatherData structure stores current and predicted weather data for the next 9 days.
+//
+// Functionality: Handler, get
+type WeatherData struct {
+	Updated string `json:"updated"`
+	Timeseries map[string]Timeseries `json:"timeseries"`
 }
 
 // Handler will handle http request for REST service.
@@ -62,7 +67,7 @@ func (weatherData *WeatherData) Handler(lat string, lon string) (int, error) {
 		//send data to database
 		var data db.Data
 		data.Container = weatherData
-		date, err := db.DB.Add("WeatherData", id, data)
+		date, _, err := db.DB.Add("WeatherData", id, data)
 		if err != nil {
 			return http.StatusInternalServerError, err
 		}
@@ -79,86 +84,66 @@ func (weatherData *WeatherData) get(lat string, lon string) (int, error) {
 	if err != nil {
 		return status, err
 	}
-	//set data in structure
-	weatherData.Now.AirTemperature = yr.Properties.Timeseries[0].Data.Instant.Details.AirTemperature
-	weatherData.Now.CloudAreaFraction = yr.Properties.Timeseries[0].Data.Instant.Details.CloudAreaFraction
-	weatherData.Now.DewPointTemperature = yr.Properties.Timeseries[0].Data.Instant.Details.DewPointTemperature
-	weatherData.Now.RelativeHumidity = yr.Properties.Timeseries[0].Data.Instant.Details.RelativeHumidity
-	weatherData.Now.WindFromDirection = yr.Properties.Timeseries[0].Data.Instant.Details.WindFromDirection
-	weatherData.Now.WindSpeed = yr.Properties.Timeseries[0].Data.Instant.Details.WindSpeed
-	weatherData.Now.WindSpeedOfGust = yr.Properties.Timeseries[0].Data.Instant.Details.WindSpeedOfGust
-	weatherData.Now.PrecipitationAmount = yr.Properties.Timeseries[0].Data.Next1Hours.Details.PrecipitationAmount
-	
-	weatherData.Today.Summary = yr.Properties.Timeseries[0].Data.Next12Hours.Summary.SymbolCode
-	weatherData.Today.Confidence = yr.Properties.Timeseries[0].Data.Next12Hours.Summary.SymbolConfidence
-	weatherData.Today.AirTemperatureMax = yr.Properties.Timeseries[0].Data.Next6Hours.Details.AirTemperatureMax
-	weatherData.Today.AirTemperatureMin = yr.Properties.Timeseries[0].Data.Next6Hours.Details.AirTemperatureMin
-	weatherData.Today.PrecipitationAmount = yr.Properties.Timeseries[0].Data.Next6Hours.Details.PrecipitationAmount
-	weatherData.Today.PrecipitationAmountMax = yr.Properties.Timeseries[0].Data.Next6Hours.Details.PrecipitationAmountMax
-	weatherData.Today.PrecipitationAmountMin = yr.Properties.Timeseries[0].Data.Next6Hours.Details.PrecipitationAmountMin
-	weatherData.Today.ProbabilityOfPrecipitation = yr.Properties.Timeseries[0].Data.Next12Hours.Details.ProbabilityOfPrecipitation
+	weatherData.Timeseries = make(map[string]Timeseries)
+	prevDate := ""
+	for _, elem := range yr.Properties.Timeseries {
+		var timeEntry Timeseries
+		rawDate, _ := time.Parse("2006-01-02T15:04:05Z", elem.Time)
+		date := rawDate.Format("2006-01-02")
+		if prevDate != date {
+			//set data in structure
+			timeEntry.Instant.AirTemperature = elem.Data.Instant.Details.AirTemperature
+			timeEntry.Instant.CloudAreaFraction = elem.Data.Instant.Details.CloudAreaFraction
+			timeEntry.Instant.DewPointTemperature = elem.Data.Instant.Details.DewPointTemperature
+			timeEntry.Instant.RelativeHumidity = elem.Data.Instant.Details.RelativeHumidity
+			timeEntry.Instant.WindFromDirection = elem.Data.Instant.Details.WindFromDirection
+			timeEntry.Instant.WindSpeed = elem.Data.Instant.Details.WindSpeed
+			timeEntry.Instant.WindSpeedOfGust = elem.Data.Instant.Details.WindSpeedOfGust
+			timeEntry.Instant.PrecipitationAmount = elem.Data.Next1Hours.Details.PrecipitationAmount
+			
+			timeEntry.Predicted.Summary = elem.Data.Next12Hours.Summary.SymbolCode
+			timeEntry.Predicted.Confidence = elem.Data.Next12Hours.Summary.SymbolConfidence
+			timeEntry.Predicted.AirTemperatureMax = elem.Data.Next6Hours.Details.AirTemperatureMax
+			timeEntry.Predicted.AirTemperatureMin = elem.Data.Next6Hours.Details.AirTemperatureMin
+			timeEntry.Predicted.PrecipitationAmount = elem.Data.Next6Hours.Details.PrecipitationAmount
+			timeEntry.Predicted.PrecipitationAmountMax = elem.Data.Next6Hours.Details.PrecipitationAmountMax
+			timeEntry.Predicted.PrecipitationAmountMin = elem.Data.Next6Hours.Details.PrecipitationAmountMin
+			timeEntry.Predicted.ProbabilityOfPrecipitation = elem.Data.Next12Hours.Details.ProbabilityOfPrecipitation
+			weatherData.Timeseries[date] = timeEntry
+			prevDate = date
+		}
+	}
 	return http.StatusOK, nil
 }
 
 func (weatherData *WeatherData) readData(data interface{}) error {
-    m := data.(map[string]interface{})
-    if data, ok := m["Now"]; ok {
-		now := data.(map[string]interface{})
-		if field, ok := now["AirTemperature"].(float64); ok {
-			weatherData.Now.AirTemperature = field
-		}
-		if field, ok := now["CloudAreaFraction"].(float64); ok {
-			weatherData.Now.CloudAreaFraction = field
-		}
-		if field, ok := now["DewPointTemperature"].(float64); ok {
-			weatherData.Now.DewPointTemperature = field
-		}
-		if field, ok := now["RelativeHumidity"].(float64); ok {
-			weatherData.Now.RelativeHumidity = field
-		}
-		if field, ok := now["WindFromDirection"].(float64); ok {
-			weatherData.Now.WindFromDirection = field
-		}
-		if field, ok := now["WindSpeed"].(float64); ok {
-			weatherData.Now.WindSpeed = field
-		}
-		if field, ok := now["WindSpeedOfGust"].(float64); ok {
-			weatherData.Now.WindSpeedOfGust = field
-		}
-		if field, ok := now["PrecipitationAmount"].(float64); ok {
-			weatherData.Now.PrecipitationAmount = field
-		}
-    } else {
-		return errors.New("getting data from database: Can't find expected fields")
-	}
-    if data, ok := m["Today"]; ok {
-		today := data.(map[string]interface{})
-		if field, ok := today["Summary"].(string); ok {
-			weatherData.Today.Summary = field
-		}
-		if field, ok := today["Confidence"].(string); ok {
-			weatherData.Today.Confidence = field
-		}
-		if field, ok := today["AirTemperatureMax"].(float64); ok {
-			weatherData.Today.AirTemperatureMax = field
-		}
-		if field, ok := today["AirTemperatureMin"].(float64); ok {
-			weatherData.Today.AirTemperatureMin = field
-		}
-		if field, ok := today["PrecipitationAmount"].(float64); ok {
-			weatherData.Today.PrecipitationAmount = field
-		}
-		if field, ok := today["PrecipitationAmountMax"].(float64); ok {
-			weatherData.Today.PrecipitationAmountMax = field
-		}
-		if field, ok := today["PrecipitationAmountMin"].(float64); ok {
-			weatherData.Today.PrecipitationAmountMin = field
-		}
-		if field, ok := today["ProbabilityOfPrecipitation"].(float64); ok {
-			weatherData.Today.ProbabilityOfPrecipitation = field
-		}
-    } else {
-		return errors.New("getting data from database: Can't find expected fields")
+	weatherData.Timeseries = make(map[string]Timeseries)
+    rawData := data.(map[string]interface{})
+	timeseries := rawData["Timeseries"].(map[string]interface{})
+	for key, elem := range timeseries {
+		var timeEntry Timeseries
+		data := elem.(map[string]interface{})
+
+		instant := data["Instant"].(map[string]interface{})
+		timeEntry.Instant.AirTemperature = instant["AirTemperature"].(float64)
+		timeEntry.Instant.CloudAreaFraction = instant["CloudAreaFraction"].(float64)
+		timeEntry.Instant.DewPointTemperature = instant["DewPointTemperature"].(float64)
+		timeEntry.Instant.RelativeHumidity = instant["RelativeHumidity"].(float64)
+		timeEntry.Instant.WindFromDirection = instant["WindFromDirection"].(float64)
+		timeEntry.Instant.WindSpeed = instant["WindSpeed"].(float64)
+		timeEntry.Instant.WindSpeedOfGust = instant["WindSpeedOfGust"].(float64)
+		timeEntry.Instant.PrecipitationAmount = instant["PrecipitationAmount"].(float64)
+
+		predicted := data["Predicted"].(map[string]interface{})
+		timeEntry.Predicted.Summary = predicted["Summary"].(string)
+		timeEntry.Predicted.Confidence = predicted["Confidence"].(string)
+		timeEntry.Predicted.AirTemperatureMax = predicted["AirTemperatureMax"].(float64)
+		timeEntry.Predicted.AirTemperatureMin = predicted["AirTemperatureMin"].(float64)
+		timeEntry.Predicted.PrecipitationAmount = predicted["PrecipitationAmount"].(float64)
+		timeEntry.Predicted.PrecipitationAmountMax = predicted["PrecipitationAmountMax"].(float64)
+		timeEntry.Predicted.PrecipitationAmountMin = predicted["PrecipitationAmountMin"].(float64)
+		timeEntry.Predicted.ProbabilityOfPrecipitation = predicted["ProbabilityOfPrecipitation"].(float64)
+		weatherData.Timeseries[key] = timeEntry
 	}
 	return nil
 }